rk_headset: avoid reporting false hook key events
The adc value may drop suddenly and then recover due to some reasons.
For example, when playback begin or unplug in a headphone. It triggers
a false hook down event.

To avoid the issue, only report a hook down event if it lasts over
a threshold. Currently the threshold is set to 200ms. This can avoid
most false events I've tested.

Change-Id: Ieeb9414655a525e54db8e80e2beec0dbdb5fb33c