media: i2c: ov8858: add s_power callback
Since commit 595e06b2d3fc ("media: i2c: fix device usage_count") the PM
core will power off the device when stream off. It results in power off
and then on when taking pictures since rkisp1 will do stream off and on
to switch the mode. It won't be a problem to ov8858 itself but an issue
to the VCM since it shares the same power as ov8858.

To avoid the issue, use the s_power callback to power on or off the
device. It holds the usage_count before s_power(on=0) is called so
stream off won't cause power off.

Change-Id: I59039f4365ca0cea3c08eb0aaeffd90dca8e3fc6